+++ src/components/custom-field-input-text/custom-field-input-text.jsx
@@ -2,19 +2,12 @@ import PropTypes from 'prop-types';
 import React, { useEffect, useRef, useState } from 'react';
 
 import cautionSvg from '../../svgs/icon-caution-fill.svg';
+import FormControl from '../form-control/form-control.jsx';
 import Icon from '../icon/icon.jsx';
 import styles from './custom-field-input-text.css';
 
-function getRootClassName(className, error, disabled) {
-  if (disabled) {
-    return `${className} ${styles.disabled}`;
-  }
-
-  if (error) {
-    return `${className} ${styles.error}`;
-  }
-
-  return className;
+function isInvalid(error, readOnly) {
+  return error && !readOnly;
 }
 
 export default function CustomFieldInputText(props) {
@@ -26,7 +19,7 @@ export default function CustomFieldInputText(props) {
   useEffect(() => {
     if (!inputRef.current) return;
 
-    if (props.error) {
+    if (isInvalid(props.error, props.readOnly)) {
       if (props.helpText) {
         inputRef.current.setCustomValidity(props.helpText);
         setValidationMessage(props.helpText);
@@ -40,7 +33,7 @@ export default function CustomFieldInputText(props) {
   });
 
   const icon = () => {
-    if (props.error) {
+    if (isInvalid(props.error, props.readOnly)) {
       return (<Icon
         className={styles['input-icon']}
         currentColor="caution"
@@ -57,41 +50,39 @@ export default function CustomFieldInputText(props) {
   };
 
   const showIcon = () => {
-    return props.error || !!props.icon;
+    return isInvalid(props.error, props.readOnly) || !!props.icon;
   };
 
-  const rootClassName = getRootClassName(props.className, props.error, props.disabled);
-
   return (
-    <div className={rootClassName} data-testid="custom-field-input" >
-      <div>
-        <label className={styles.label} htmlFor={props.id}>{props.label}</label>
-        {props.required && <span className={styles.required}>(Required)</span>}
-      </div>
-      <div className={styles['input-container']}>
-        <input
-          defaultValue={props.value}
-          className={styles.input}
-          disabled={props.disabled}
-          id={props.id}
-          max={props.max}
-          min={props.min}
-          name={props.name}
-          onBlur={props.onBlur}
-          onChange={props.onChange}
-          onFocus={props.onFocus}
-          onKeyUp={props.onKeyUp}
-          placeholder={props.placeholder}
-          readOnly={props.readOnly}
-          ref={inputRef}
-          required={props.required}
-          step={props.step}
-          type={props.type}
-        />
-        {showIcon() && icon()}
-      </div>
-      {props.error && <span className={styles.help}>{validationMessage}</span>}
-    </div>
+    <FormControl
+      className={props.className}
+      error={validationMessage}
+      id={props.id}
+      label={props.label}
+      readOnly={props.readOnly}
+      required={props.required}
+    >
+      <input
+        defaultValue={props.value}
+        className={styles.input}
+        disabled={props.disabled}
+        id={props.id}
+        max={props.max}
+        min={props.min}
+        name={props.name}
+        onBlur={props.onBlur}
+        onChange={props.onChange}
+        onFocus={props.onFocus}
+        onKeyUp={props.onKeyUp}
+        placeholder={props.placeholder}
+        readOnly={props.readOnly}
+        ref={inputRef}
+        required={props.required}
+        step={props.step}
+        type={props.type}
+      />
+      {showIcon() && icon()}
+    </FormControl>
   );
 }
 
@@ -133,7 +124,7 @@ CustomFieldInputText.propTypes = {
 };
 
 CustomFieldInputText.defaultProps = {
-  className: styles['custom-field-input-text'],
+  className: undefined,
   disabled: false,
   error: false,
   helpText: undefined,